py-otl/pyproject.toml
[tool.poetry]
name = "pyotl"
version = "0.1.0"
description = "otl is a system to describe tests for eda"
authors = ["James <james@silogy.io>"]


[tool.poetry.dependencies]
python = "^3.8"
typer = "0.9.0"
pyyaml = "6.0.1"
pydantic = "2.6.4"
toml = "0.10.2"
betterproto = { version = "1.2.5", extras = ["compiler"] }
rich = "^13.7.1"


[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
pytest = "^7.4.3"
sphinx = "7.1.0"
setuptools = "^69.2.0"


#we need this for maturin 
[project]
name = "pyotl"
dependencies = [
  "typer>=0.9.0",
  "pyyaml>=6.0.1",
  "pydantic>=2.6.0",
  "toml>=0.10.2",
  "betterproto[compiler]>=1.2.5",
  "rich>=13.7.1",

]


[tool.poetry.scripts]
otl-cli = "otl.cli:main"


[build-system]
requires = ["maturin>=1.5,<2.0"]
build-backend = "maturin"


[tool.maturin]
features = ["pyo3/extension-module"]
